Break out GetBlockCapacityByte for linux and unsupported platforms
GetBlockCapacityByte for contains reference to a constant unix.BLKGETSIZE64
which is not defined for darwin.

This PR breaks out the GetBlockCapacityByte method for linux and unsupported,
platforms which enables make test and development activities on darwin (aka mac)
to PASS.
